django_app/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-celery-beat>=2.6,<3.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
anthropic>=0.40,<1.0
reportlab>=4.1,<5.0
requests>=2.31,<3.0
whitenoise>=6.6,<7.0
